debugging

+++ streamlit_app.py
@@ -80,8 +80,8 @@ st.success("Data loaded successfully! Please use the visualization options below
 
 
 year_filter = st.slider(label = 'Year', 
-                        min_value = df['Year'].min(), 
-                        max_value = df['Year'].max(),
+                        min_value = np.int64(df['Year'].min()), 
+                        max_value = np.int64(df['Year'].max()),
                         value = (2007,2009), step = np.int64(1),
                         help="Select the year range.")
